requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2

# Database and utilities
python-dateutil==2.8.2
requests==2.31.0
psycopg2-binary==2.9.10

# Development tools
python-dotenv==1.0.0

# AI/LLM Integration
google-generativeai==0.8.3
